+++ src/components/mypage/mobile/MobileMyPage.tsx
@@ -1,15 +1,17 @@
-import styled from '@emotion/styled';
 import { useState } from 'react';
+import styled from '@emotion/styled';
+import MobileTopTab from './MobileTopTab';
 import MobileUserInfo from './MobileUserInfo';
-import TobTab from './TobTab';
+import { ProjectListWrapper } from 'pages/MyPage';
 
 const MobileMyPage = ({ user }: MypageInfoProps) => {
   const [activeTab, setActiveTab] = useState('개인정보');
 
   return (
     <MobileMyPageContainer>
-      <TobTab activeTab={activeTab} setActiveTab={setActiveTab} />
+      <MobileTopTab activeTab={activeTab} setActiveTab={setActiveTab} />
       {activeTab === '개인정보' && <MobileUserInfo user={user} />}
+      {activeTab === '프로젝트' && <ProjectListWrapper></ProjectListWrapper>}
     </MobileMyPageContainer>
   );
 };
